Se incluyen modificaciones en UsuariosSancionados, y se crea y termina usuarios con las mismas modificaciones de manejo de excepciones.
Queda pendiente arreglar Estadios y Estadios Secciones, luego elegir los puertos de c/u y conectarlos con el revisador.

+++ ticketgol/usuarios_sancionados/src/main/java/ticketgol/usuarios_sancionados/service/UsuarioSancionadoService.java
@@ -1,5 +1,7 @@
 package ticketgol.usuarios_sancionados.service;
 
+import ticketgol.usuarios_sancionados.exception.UsuarioAlreadyExistsException;
+import ticketgol.usuarios_sancionados.exception.UsuarioNotFoundException;
 import ticketgol.usuarios_sancionados.model.UsuarioSancionado;
 import ticketgol.usuarios_sancionados.repository.UsuarioSancionadoRepository;
 import jakarta.transaction.Transactional;
@@ -21,14 +23,45 @@ public class UsuarioSancionadoService {
 
     public UsuarioSancionado findById(Long id) {
         return usuarioSancionadoRepository.findById(id)
-                .orElseThrow(() -> new RuntimeException("Usuario sancionado no encontrado"));
+                .orElseThrow(() ->
+                        new UsuarioNotFoundException(
+                                "Usuario sancionado con ID " + id + " no encontrado"
+                        )
+                );
     }
 
     public UsuarioSancionado save(UsuarioSancionado usuarioSancionado) {
+
+        if (usuarioSancionadoRepository.existsByRut(usuarioSancionado.getRut())) {
+            throw new UsuarioAlreadyExistsException(
+                    "Ya existe un usuario sancionado con RUT " +
+                            usuarioSancionado.getRut()
+            );
+        }
+
         return usuarioSancionadoRepository.save(usuarioSancionado);
     }
 
+    public UsuarioSancionado update(Long id, UsuarioSancionado usuarioActualizado) {
+
+        UsuarioSancionado usuarioExistente = findById(id);
+
+        usuarioExistente.setRut(usuarioActualizado.getRut());
+        usuarioExistente.setMotivo(usuarioActualizado.getMotivo());
+        usuarioExistente.setFechaSancion(usuarioActualizado.getFechaSancion());
+        usuarioExistente.setFechaExpiracion(usuarioActualizado.getFechaExpiracion());
+
+        return usuarioSancionadoRepository.save(usuarioExistente);
+    }
+
     public void delete(Long id) {
+
+        if (!usuarioSancionadoRepository.existsById(id)) {
+            throw new UsuarioNotFoundException(
+                    "Usuario sancionado con ID " + id + " no encontrado"
+            );
+        }
+
         usuarioSancionadoRepository.deleteById(id);
     }
 }
